fixed last bug
